refactor: consolidate duplicated getDomain() and applyTheme()
- Move applyTheme() from popup.ts and options.ts to new shared/theme.ts
- Replace popup.ts local getDomain() with import from shared/domain.ts
- Update popup.test.ts and options.test.ts to import from shared modules
- Functions extracted as-is, no logic changes

Closes #12
